# vectispire-angular/openapi.json
{
  "components" : {
    "schemas" : {
...
      "Builder" : {
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "version" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ComplianceControl" : {
        "properties" : {
          "category" : {
            "enum" : [ "VULNERABILITY_MANAGEMENT", "SUPPLY_CHAIN", "SECRETS_MANAGEMENT", "SECURE_CODING", "INFRASTRUCTURE_AS_CODE", "GOVERNANCE", "AUDIT_AND_LOGGING" ],
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "requirement" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ComplianceEvaluation" : {
        "properties" : {
          "controls" : {
            "items" : {
              "$ref" : "#/components/schemas/ControlAssessment"
            },
            "type" : "array"
          },
          "framework" : {
            "enum" : [ "NIS_2", "ISO_27001", "EU_CRA", "DORA", "PCI_DSS", "SOC_2" ],
            "type" : "string"
          },
          "overallStatus" : {
            "enum" : [ "COMPLIANT", "PARTIAL", "NON_COMPLIANT" ],
            "type" : "string"
          },
          "scorePercentage" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "ComplianceSummary" : {
        "properties" : {
          "dueSoonCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "evaluations" : {
            "items" : {
              "$ref" : "#/components/schemas/ComplianceEvaluation"
            },
            "type" : "array"
          },
          "mttr" : {
            "$ref" : "#/components/schemas/MttrResult"
          },
          "overdueCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "passingGateTargets" : {
            "format" : "int32",
            "type" : "integer"
          },
          "targets" : {
            "items" : {
              "$ref" : "#/components/schemas/TargetCompliance"
            },
            "type" : "array"
          },
          "totalMonitoredTargets" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ControlAssessment" : {
        "properties" : {
          "control" : {
            "$ref" : "#/components/schemas/ComplianceControl"
          },
          "details" : {
            "type" : "string"
          },
          "remediationGuidance" : {
            "type" : "string"
          },
          "scorePercentage" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "COMPLIANT", "PARTIAL", "NON_COMPLIANT" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "FindingView" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "filePath" : {
            "type" : "string"
          },
          "fixVersions" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "identifier" : {
            "type" : "string"
          },
          "line" : {
            "format" : "int32",
            "type" : "integer"
          },
          "link" : {
            "type" : "string"
          },
          "packageName" : {
            "type" : "string"
          },
          "packageVersion" : {
            "type" : "string"
          },
          "severity" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "FindingsSummary" : {
        "properties" : {
          "critical" : {
            "format" : "int64",
            "type" : "integer"
          },
          "high" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kev" : {
            "format" : "int64",
            "type" : "integer"
          },
          "low" : {
            "format" : "int64",
            "type" : "integer"
          },
          "medium" : {
            "format" : "int64",
            "type" : "integer"
          },
          "secrets" : {
            "format" : "int64",
            "type" : "integer"
          },
          "total" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "InTotoAttestation" : {
        "properties" : {
          "_type" : {
            "type" : "string"
          },
          "predicate" : {
            "$ref" : "#/components/schemas/Predicate"
          },
          "predicateType" : {
            "type" : "string"
          },
          "subject" : {
            "items" : {
              "$ref" : "#/components/schemas/Subject"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "Invocation" : {
        "properties" : {
          "branch" : {
            "type" : "string"
          },
          "commitSha" : {
            "type" : "string"
          },
          "scanId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "targetKind" : {
            "type" : "string"
          },
          "targetName" : {
            "type" : "string"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "MttrResult" : {
        "properties" : {
          "mttrBySeverityDays" : {
            "additionalProperties" : {
              "format" : "double",
              "type" : "number"
            },
            "type" : "object"
          },
          "overallMttrDays" : {
            "format" : "double",
            "type" : "number"
          },
          "resolvedCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PolicyAssessment" : {
        "properties" : {
          "enforcedPolicy" : {
            "type" : "string"
          },
          "gatePassed" : {
            "type" : "boolean"
          },
          "violations" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "Predicate" : {
        "properties" : {
          "builder" : {
            "$ref" : "#/components/schemas/Builder"
          },
          "findings" : {
            "$ref" : "#/components/schemas/FindingsSummary"
          },
          "invocation" : {
            "$ref" : "#/components/schemas/Invocation"
          },
          "policy" : {
            "$ref" : "#/components/schemas/PolicyAssessment"
          },
          "sbomDigestSha256" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ScanDetail" : {
        "properties" : {
          "findings" : {
            "items" : {
              "$ref" : "#/components/schemas/FindingView"
            },
            "type" : "array"
          },
          "findingsTotal" : {
            "format" : "int64",
            "type" : "integer"
          },
          "findingsTruncated" : {
            "type" : "boolean"
          },
          "hasSbom" : {
            "type" : "boolean"
          },
          "projectType" : {
            "type" : "string"
          },
          "projectVersion" : {
            "type" : "string"
          },
          "scan" : {
            "$ref" : "#/components/schemas/ScanSummary"
          },
          "subPath" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ScanSummary" : {
        "properties" : {
          "attempts" : {
            "format" : "int32",
            "type" : "integer"
          },
          "branch" : {
            "type" : "string"
          },
          "claimedBy" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "durationMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "error" : {
            "type" : "string"
          },
          "findingsCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "newIssuesCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "resolvedIssuesCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "type" : "string"
          },
          "targetId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "targetKind" : {
            "type" : "string"
          },
          "targetName" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Subject" : {
        "properties" : {
          "digest" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "type" : "object"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TargetCompliance" : {
        "properties" : {
          "frameworkScores" : {
            "additionalProperties" : {
              "format" : "int32",
              "type" : "integer"
            },
            "type" : "object"
          },
          "gateStatus" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "openIssuesCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "overallScore" : {
            "format" : "int32",
            "type" : "integer"
          },
          "overallStatus" : {
            "type" : "string"
          },
          "overdueCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "targetId" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TeamTargetAssignment" : {
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UserTargetAssignment" : {
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "AgentKeyAuth" : {
        "description" : "Pre-shared authentication key for remote scanning agents",
        "in" : "header",
        "name" : "X-Agent-Key",
        "type" : "apiKey"
      },
      "ApiKeyAuth" : {
        "description" : "Programmatic API Key for CI/CD and automation integrations",
        "in" : "header",
        "name" : "X-API-Key",
        "type" : "apiKey"
      },
      "BearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "JWT Session Bearer token obtained via /api/v1/auth/login",
        "name" : "BearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "name" : "Asmolabs Security Team",
      "url" : "https://github.com/asmolabs/vectispire"
    },
    "description" : "Automated supply-chain security, SBOM analysis, vulnerability triage, attack surface inventory, and compliance governance platform.",
    "license" : {
      "name" : "Apache-2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "Vectispire Control Plane REST API",
    "version" : "4.1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/attestations/scans/{scanId}" : {
      "get" : {
        "operationId" : "forScan",
        "parameters" : [ {
          "in" : "path",
          "name" : "scanId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InTotoAttestation"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "attestation-controller" ]
      }
    },
...
    "/api/v1/compliance/summary" : {
      "get" : {
        "description" : "Returns compliance scores across all regulatory frameworks.",
        "operationId" : "summary",
        "parameters" : [ {
          "description" : "Optional target ID filter",
          "in" : "query",
          "name" : "targetId",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ComplianceSummary"
                }
              }
            },
            "description" : "Compliance summary evaluated successfully"
          }
        },
        "summary" : "Get compliance summary",
        "tags" : [ "Compliance" ]
      }
    },
...
    "/api/v1/scans" : {
      "get" : {
        "description" : "Returns historical security scans with filtering by repository or container target.",
        "operationId" : "list_10",
        "parameters" : [ {
          "description" : "Filter by repository ID",
          "in" : "query",
          "name" : "repo_id",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "Filter by container ID",
          "in" : "query",
          "name" : "container_id",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "Maximum results to return",
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ScanSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Scan history retrieved successfully"
          }
        },
        "summary" : "List scan history",
        "tags" : [ "Scans" ]
      }
    },
    "/api/v1/scans/{id}" : {
      "get" : {
        "description" : "Returns full details and raw findings observed during a specific scan.",
        "operationId" : "detail",
        "parameters" : [ {
          "description" : "Scan ID",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ScanDetail"
                }
              }
            },
            "description" : "Scan details retrieved successfully"
          }
        },
        "summary" : "Get scan detail",
        "tags" : [ "Scans" ]
      }
    },
    "/api/v1/scans/{id}/sbom" : {
      "get" : {
        "description" : "Returns the complete Software Bill of Materials (SBOM) produced during this scan.",
        "operationId" : "sbom",
        "parameters" : [ {
          "description" : "Scan ID",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Syft native JSON SBOM document, exactly as the cataloguer produced it"
          }
        },
        "summary" : "Download scan SBOM",
        "tags" : [ "Scans" ]
      }
    },
...
    "/api/v1/teams/{id}/targets" : {
      "get" : {
        "operationId" : "targets_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TeamTargetAssignment"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "teams-controller" ]
      },
      "put" : {
        "operationId" : "setTargets_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "items" : {
                  "$ref" : "#/components/schemas/TeamTargetAssignment"
                },
                "type" : "array"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TeamTargetAssignment"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "teams-controller" ]
      }
    },
...
    "/api/v1/users/{id}/targets" : {
      "get" : {
        "operationId" : "targets",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/UserTargetAssignment"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "users-controller" ]
      },
      "put" : {
        "operationId" : "setTargets",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "items" : {
                  "$ref" : "#/components/schemas/UserTargetAssignment"
                },
                "type" : "array"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/UserTargetAssignment"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "users-controller" ]
      }
    },
...
  },
  "security" : [ {
    "BearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Current Vectispire Instance",
    "url" : "/"
  } ],
...
}